Fix some require cycles (#1383)
### Description

Fix some require cycles which were causing issues for the work I'm doing for #305
- between account/reducer and backup/utils
- src/identity/reducer.ts -> src/identity/verification.ts -> src/app/selectors.ts -> src/identity/reducer.ts

Require cycles can create subtle issues, we should aim to remove all of them to avoid surprises in the future.

Before this change, the RN packager was listing 89 require cycle warnings.
After this change, we're down to 80 require cycle warnings.
