+++ packages/mobile/src/account/profileInfo.ts
@@ -9,7 +9,8 @@ import { call, put, select } from 'redux-saga/effects'
 import { profileUploaded } from 'src/account/actions'
 import { isProfileUploadedSelector, nameSelector, pictureSelector } from 'src/account/selectors'
 import UploadServiceDataWrapper from 'src/account/UploadServiceDataWrapper'
-import { walletToAccountAddressSelector, WalletToAccountAddressType } from 'src/identity/reducer'
+import { WalletToAccountAddressType } from 'src/identity/reducer'
+import { walletToAccountAddressSelector } from 'src/identity/selectors'
 import { DEK, retrieveOrGeneratePepper } from 'src/pincode/authentication'
 import { extensionToMimeType, getDataURL, saveRecipientPicture } from 'src/utils/image'
 import Logger from 'src/utils/Logger'
