+++ packages/mobile/src/app/verificationMigration.test.ts
@@ -4,7 +4,7 @@ import { call, select } from 'redux-saga-test-plan/matchers'
 import { verificationMigrationRan } from 'src/app/actions'
 import { numberVerifiedSelector, ranVerificationMigrationSelector } from 'src/app/selectors'
 import { runVerificationMigration } from 'src/app/verificationMigration'
-import { e164NumberToSaltSelector } from 'src/identity/reducer'
+import { e164NumberToSaltSelector } from 'src/identity/selectors'
 import { e164NumberSelector, komenciContextSelector } from 'src/verify/reducer'
 import { getContractKitAsync } from 'src/web3/contracts'
 import { mtwAddressSelector } from 'src/web3/selectors'
