+++ packages/mobile/src/app/verificationMigration.ts
@@ -5,7 +5,8 @@ import { getPhoneHash } from '@celo/utils/lib/phoneNumbers'
 import { call, put, select } from 'redux-saga-test-plan/matchers'
 import { verificationMigrationRan } from 'src/app/actions'
 import { numberVerifiedSelector, ranVerificationMigrationSelector } from 'src/app/selectors'
-import { e164NumberToSaltSelector, E164NumberToSaltType } from 'src/identity/reducer'
+import { E164NumberToSaltType } from 'src/identity/reducer'
+import { e164NumberToSaltSelector } from 'src/identity/selectors'
 import Logger from 'src/utils/Logger'
 import { e164NumberSelector, KomenciContext, komenciContextSelector } from 'src/verify/reducer'
 import { getContractKit } from 'src/web3/contracts'
