+++ packages/mobile/src/backup/consts.ts
@@ -1,0 +1,2 @@
+export const DAYS_TO_BACKUP = 1
+export const DAYS_TO_DELAY = 1 / 24 // 1 hour delay
