+++ packages/mobile/src/backup/utils.ts
@@ -13,11 +13,7 @@ import { currentAccountSelector } from 'src/web3/selectors'
 
 const TAG = 'Backup/utils'
 
-export const DAYS_TO_BACKUP = 1
-export const DAYS_TO_DELAY = 1 / 24 // 1 hour delay
-export const MNEMONIC_SPLITTER = 'celo'
-
-export const MNEMONIC_STORAGE_KEY = 'mnemonic'
+const MNEMONIC_STORAGE_KEY = 'mnemonic'
 
 export function getMnemonicLanguage(language: string | null) {
   switch (language?.slice(0, 2)) {
