+++ packages/mobile/src/home/NotificationBox.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render } from '@testing-library/react-native'
 import * as React from 'react'
 import { Provider } from 'react-redux'
 import { openUrl } from 'src/app/actions'
-import { DAYS_TO_BACKUP } from 'src/backup/utils'
+import { DAYS_TO_BACKUP } from 'src/backup/consts'
 import NotificationBox from 'src/home/NotificationBox'
 import { Currency } from 'src/utils/currencies'
 import { createMockStore, getElementText } from 'test/utils'
