+++ packages/mobile/src/identity/commentEncryption.ts
@@ -25,11 +25,10 @@ import {
 import { getUserSelfPhoneHashDetails } from 'src/identity/privateHashing'
 import {
   AddressToE164NumberType,
-  e164NumberToAddressSelector,
   E164NumberToAddressType,
-  e164NumberToSaltSelector,
   E164NumberToSaltType,
 } from 'src/identity/reducer'
+import { e164NumberToAddressSelector, e164NumberToSaltSelector } from 'src/identity/selectors'
 import { NewTransactionsInFeedAction } from 'src/transactions/actions'
 import Logger from 'src/utils/Logger'
 import { getContractKit } from 'src/web3/contracts'
