+++ packages/mobile/src/identity/saga.ts
@@ -18,9 +18,10 @@ import {
 } from 'src/identity/actions'
 import { checkTxsForIdentityMetadata } from 'src/identity/commentEncryption'
 import { doImportContactsWrapper, fetchAddressesAndValidateSaga } from 'src/identity/contactMapping'
-import { AddressValidationType, e164NumberToAddressSelector } from 'src/identity/reducer'
+import { AddressValidationType } from 'src/identity/reducer'
 import { revokeVerificationSaga } from 'src/identity/revoke'
 import { validateAndReturnMatch } from 'src/identity/secureSend'
+import { e164NumberToAddressSelector } from 'src/identity/selectors'
 import { reportRevealStatusSaga, startVerificationSaga } from 'src/identity/verification'
 import { recipientHasNumber } from 'src/recipients/recipient'
 import { Actions as TransactionActions } from 'src/transactions/actions'
