+++ packages/mobile/src/paymentRequest/OutgoingPaymentRequestListScreen.tsx
@@ -3,7 +3,8 @@ import { WithTranslation } from 'react-i18next'
 import { View } from 'react-native'
 import { connect } from 'react-redux'
 import i18n, { Namespaces, withTranslation } from 'src/i18n'
-import { e164NumberToAddressSelector, E164NumberToAddressType } from 'src/identity/reducer'
+import { E164NumberToAddressType } from 'src/identity/reducer'
+import { e164NumberToAddressSelector } from 'src/identity/selectors'
 import {
   NotificationList,
   titleWithBalanceNavigationOptions,
