+++ packages/mobile/src/send/SendAmount/useTransactionCallbacks.ts
@@ -9,8 +9,9 @@ import { TokenTransactionType } from 'src/apollo/types'
 import { ErrorMessages } from 'src/app/ErrorMessages'
 import { STABLE_TRANSACTION_MIN_AMOUNT } from 'src/config'
 import { getFeeEstimateDollars } from 'src/fees/selectors'
-import { AddressValidationType, secureSendPhoneNumberMappingSelector } from 'src/identity/reducer'
+import { AddressValidationType } from 'src/identity/reducer'
 import { getAddressValidationType } from 'src/identity/secureSend'
+import { secureSendPhoneNumberMappingSelector } from 'src/identity/selectors'
 import { RecipientVerificationStatus } from 'src/identity/types'
 import { convertDollarsToLocalAmount } from 'src/localCurrency/convert'
 import {
