+++ packages/mobile/src/send/saga.ts
@@ -11,7 +11,7 @@ import { ErrorMessages } from 'src/app/ErrorMessages'
 import { calculateFee, FeeInfo } from 'src/fees/saga'
 import { transferGoldToken } from 'src/goldToken/actions'
 import { encryptComment } from 'src/identity/commentEncryption'
-import { e164NumberToAddressSelector } from 'src/identity/reducer'
+import { e164NumberToAddressSelector } from 'src/identity/selectors'
 import { sendInvite } from 'src/invite/saga'
 import { navigateBack, navigateHome } from 'src/navigator/NavigationService'
 import { completePaymentRequest } from 'src/paymentRequest/actions'
