+++ packages/mobile/src/transactions/CeloTransferFeedItem.tsx
@@ -16,7 +16,7 @@ import CurrencyDisplay from 'src/components/CurrencyDisplay'
 import { formatShortenedAddress } from 'src/components/ShortenedAddress'
 import { txHashToFeedInfoSelector } from 'src/fiatExchanges/reducer'
 import { Namespaces } from 'src/i18n'
-import { addressToDisplayNameSelector } from 'src/identity/reducer'
+import { addressToDisplayNameSelector } from 'src/identity/selectors'
 import { getRecipientFromAddress } from 'src/recipients/recipient'
 import { recipientInfoSelector } from 'src/recipients/reducer'
 import { navigateToPaymentTransferReview } from 'src/transactions/actions'
