+++ packages/mobile/src/transactions/TransferConfirmationCard.tsx
@@ -21,7 +21,7 @@ import TotalLineItem from 'src/components/TotalLineItem'
 import { FAQ_LINK } from 'src/config'
 import { RewardsScreenOrigin } from 'src/consumerIncentives/analyticsEventsTracker'
 import { Namespaces } from 'src/i18n'
-import { addressToDisplayNameSelector } from 'src/identity/reducer'
+import { addressToDisplayNameSelector } from 'src/identity/selectors'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { Recipient } from 'src/recipients/recipient'
