+++ packages/mobile/src/transactions/TransferFeedItem.tsx
@@ -7,7 +7,8 @@ import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { TokenTransactionType, TransferItemFragment } from 'src/apollo/types'
 import { txHashToFeedInfoSelector } from 'src/fiatExchanges/reducer'
 import { Namespaces } from 'src/i18n'
-import { addressToDisplayNameSelector, AddressToE164NumberType } from 'src/identity/reducer'
+import { AddressToE164NumberType } from 'src/identity/reducer'
+import { addressToDisplayNameSelector } from 'src/identity/selectors'
 import { InviteDetails } from 'src/invite/actions'
 import { getRecipientFromAddress, NumberToRecipient, RecipientInfo } from 'src/recipients/recipient'
 import { inviteRewardsSendersSelector, rewardsSendersSelector } from 'src/recipients/reducer'
